enh: better structure for the dpmodel uts
