+++ source/tests/common/dpmodel/README
@@ -1,0 +1,1 @@
+test deepmd-kit/source/deepmd/dpmodel
