+++ source/tests/common/dpmodel/test_pairtab.py
